Write page form
Just a plain form. Of course the form's gonna be an "inline" form. I'm
not sure if it's a good idea to make the form itself the flex container
in the future, or wrap its contents in a div and make it the flex
container instead... Hmmm...

Also a section where search results will be populated. Maybe we should
have included an id now too...