Add babel polyfill for generators
The polyfill bloats our bundle (as of this commit the dev bundle is
1.7MB while the production bundle is 100kB). But because this commit
introduces generators, we have no choice but to include the polyfill
too.

This makes looking up ways of further reducing the bundle size more
important.

PS: I forgot to include .babelrc in the previous commit. Yeah, I
should amend, but I'm already typing this commit message in vim, and
I don't want to exit just to amend and then retype everything. Shame.