Write main search script
The star of this script is the use of JS generators, promises, and the
fetch API. The Wikipedia API used doesn't seem to return data that's
rejectable by the promise (is that the right term?), so there's no
opportunity to use `.catch()` in the promise chain, or wrapping the
`yield fetchWikipediaPages` line in a try-catch block.

The use of template literals is pretty slick too.

Maybe there's also a chance of splitting the file into modules.