Use spinning icon to indicate loading
Made a function that returns a plain string to indicate that a fetch is
being performed. Wrapping it in a function also keeps clutter out of the
generator code.

Now I'm loading the entire Font Awesome library for just a couple of
classes, which sucks. We should look up how to include the library in
our webpack bundle so we cut out code that we don't use.