Modify git script to deploy only on master branch
The surge command will only run if `git push` was run while on the
master branch, and if the build is successful.

I refactored `surge` in its own npm script so the git script is not
awfully long.

Note that this check is not perfect; it will attempt a deploy whenever a
push is about to be made, regardless whether the branch to be pushed is
master or not.