Include font awesome in webpack bundle
Font Awesome is now a listed dependency, so the <link> element for that
is no longer needed.

The source from which I saw how to include FA in the bundle
(www.mattzabriskie.com/blog/font-awesome-webpack) also included the
Extract Text plugin, but I don't think it makes sense to include that in
the dev config (we'll lose hot module replacement for styles!).