Apply color palette
We generated a nice looking palette using coolors.co. Overall the page
color is gray, but the background color of the results is a lighter
shade. The main heading is brown while the links to articles is a shade
of red (maybe the colors should be swapped?). The rest of the text is
almost black.