Apply styles to controls section
Aside from centering the controls, we made the form a flex container.
The search box occupies the rest of the space while the buttons have a
fixed width.

We had to touch the resuls section for this commit. In particular, I
removed the first result's top margin, since it's "leaking" into the
controls section.